backend/requirements.txt
fastapi
python-socketio
uvicorn
google-generativeai
